Migrate fjs/types/byte_set to .f.mjs, split types into types.ts
Splits the ByteSet type into a sibling types.ts, and converts the
runtime implementation to JSDoc-typed module.f.mjs/proof.f.mjs under
the stage-1 TypeScript-to-mjs migration. Updates the two dependents
(fjs/fsm/module.f.ts, fjs/fsm/proof.f.ts) to import runtime values from
module.f.mjs and the ByteSet type from types.ts, and fixes a doc link
in fjs/types/nibble_set/module.f.mjs.
