+++ fjs/fsm/module.f.ts
@@ -6,7 +6,8 @@
 import type { List } from '../types/list/types.ts'
 import { equal, isEmpty, fold, toArray, scan, foldScan, empty as emptyList } from '../types/list/module.f.mjs'
 import type { StringMap } from '../types/object/types.ts'
-import { toRangeMap, union as byteSetUnion, one, empty, range, type ByteSet } from '../types/byte_set/module.f.ts'
+import { toRangeMap, union as byteSetUnion, one, empty, range } from '../types/byte_set/module.f.mjs'
+import type { ByteSet } from '../types/byte_set/types.ts'
 import { intersect, union as sortedSetUnion } from '../types/sorted_set/module.f.mjs'
 import type { SortedSet } from '../types/sorted_set/types.ts'
 import type { RangeMap, Properties, RangeMapArray, Entry } from '../types/range_map/types.ts'
