+++ fjs/types/byte_set/proof.f.mjs
@@ -1,11 +1,10 @@
-import { has, empty, set, setRange, unset, universe, complement, toRangeMap } from './module.f.ts'
+import { has, empty, set, setRange, unset, universe, complement, toRangeMap } from './module.f.mjs'
 import { every, countdown, map, toArray } from '../list/module.f.mjs'
-import { stringify as jsonStringify, type Unknown } from '../../media/json/module.f.ts'
+import { stringify as jsonStringify } from '../../media/json/module.f.ts'
 import { sort } from '../object/module.f.mjs'
 import { assert, assertEq } from '../../asserts/module.f.mjs'
 
-const stringify: (a: readonly Unknown[]) => string
-    = jsonStringify(sort)
+const stringify = jsonStringify(sort)
 
 export const proof = {
     has: [
@@ -46,7 +45,7 @@ export const proof = {
         }
     ],
     universe: () => {
-        const x = every(map((v: any) => has(v)(universe))(countdown(256)))
+        const x = every(map(v => has(v)(universe))(countdown(256)))
         assert(x, x)
     },
     compliment: {
