+++ fjs/types/byte_set/types.ts
@@ -1,0 +1,7 @@
+/**
+ * Type-level API for the byte-set module.
+ *
+ * @module
+ */
+
+export type ByteSet = bigint
